requirements.txt
Flask
python-dotenv
PyGithub
openai